Adding verification for deleting and updating an Restaurant

+++ src/services/restaurant.service.ts
@@ -12,7 +12,10 @@ import { RestaurantDto } from '../dto/restaurant.dto';
 export class RestaurantService {
   constructor(private readonly restaurantRepository: RestaurantRepository) {}
 
-  addRestaurant = async (parameters: addRestaurantDto): Promise<Restaurant> => {
+  addRestaurant = async (
+    parameters: addRestaurantDto,
+    userId: string,
+  ): Promise<Restaurant> => {
     const { name, foodType, address, zipcode } = parameters;
 
     const restaurant = await this.restaurantRepository.insert({
@@ -20,6 +23,7 @@ export class RestaurantService {
       foodType,
       address,
       zipcode,
+      userId,
     });
 
     return restaurant.save();
@@ -33,17 +37,26 @@ export class RestaurantService {
     return this.restaurantRepository.findOneById(restaurantId);
   };
 
-  deleteRestaurantById = async (restaurantId: string) => {
-    return await this.restaurantRepository.deleteOnyBy({ _id: restaurantId });
-  };
-
   updateRestaurantById = async (
     restaurantId: string,
     parameters: RestaurantDto,
+    userId: string,
   ) => {
+    const restaurant = await this.getRestaurantByID(restaurantId);
+    if (restaurant.userId !== userId.toString()) {
+      throw new UnauthorizedException();
+    }
     await this.restaurantRepository.updateOneBy(
       { _id: restaurantId },
       { ...parameters },
     );
+  };
+
+  deleteRestaurantById = async (restaurantId: string, userId: string) => {
+    const restaurant = await this.getRestaurantByID(restaurantId);
+    if (restaurant.userId !== userId.toString()) {
+      throw new UnauthorizedException();
+    }
+    return await this.restaurantRepository.deleteOnyBy({ _id: restaurantId });
   };
 }
